Refactor the chart out of measurement settings page
